requirements.txt
# VoxTrail — Python Dependencies
# Generated for Python 3.11+

# ── Web Framework ──────────────────────────────────────────────────
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
sse-starlette>=1.8.2
python-multipart>=0.0.9

# ── Google AI / ADK ───────────────────────────────────────────────
google-generativeai>=0.8.0
google-adk>=0.5.0
google-cloud-aiplatform>=1.60.0

# ── Auth & Security ───────────────────────────────────────────────
msal>=1.28.0
pyjwt[crypto]>=2.8.0
cryptography>=42.0.0
pycryptodome>=3.20.0

# ── Data Validation ───────────────────────────────────────────────
pydantic>=2.7.0
python-dotenv>=1.0.0

# ── Redis & MCP ───────────────────────────────────────────────────
redis>=5.0.0
mcp>=1.0.0

# ── HTTP Clients ──────────────────────────────────────────────────
httpx>=0.27.0
requests>=2.31.0

# ── Database ──────────────────────────────────────────────────────
sqlalchemy>=2.0.30

# ── Observability ─────────────────────────────────────────────────
opentelemetry-api>=1.24.0
opentelemetry-sdk>=1.24.0
opentelemetry-exporter-otlp-proto-http>=1.24.0
arize-phoenix-otel>=0.3.0

# ── Logging ───────────────────────────────────────────────────────
loguru>=0.7.2

# ── Audio (voicebot/ utilities) ───────────────────────────────────
sounddevice>=0.4.6
soundfile>=0.12.1
numpy>=1.26.0
websockets>=12.0
